Move OIDC stuff from hkijwt to oidc_apis
Move the API scope and related models to a new app "oidc_apis" for
cleaner separation between the new OIDC and the old JWT implementation.

Note: The migrations will REMOVE DATA of the old models and won't
migrate any data to the new tables.
